does this even work?  I mean, it compiles on my machine so, probably not.
These changes are using main.c as the basis for a snake game.  I guess I'm just going to periodically save a copy as something else, like main.snake.c, whenever it's time to work on a different project.

That feels so janky to me, but I guess if it works for now, it can evolve into a good system later.

I couldn't get mmap to work, so I'm just using malloc.  I don't know if that actually guarantees contiguous memory or not, but it's late and I need to go to bed, so this is it for today.